player/requirements.txt
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioice==0.10.1
aiortc==1.11.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==14.3.0
black==25.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
coloredlogs==15.0.1
cryptography==44.0.3
daily-python==0.18.2
dataclasses-json==0.6.7
deepgram-sdk==3.8.0
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
flatbuffers==25.2.10
frozenlist==1.6.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-speech==2.32.0
google-cloud-texttospeech==2.27.0
google-crc32c==1.7.1
google-genai==1.15.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
iniconfig==2.1.0
jiter==0.9.0
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.8
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
numba==0.61.2
numpy==1.26.4
onnxruntime==1.20.1
openai==1.70.0
opencv-python==4.11.0.86
ormsgpack==1.7.0
packaging==25.0
pathspec==0.12.1
pillow==11.1.0
pipecat-ai==0.0.67
platformdirs==4.3.8
pluggy==1.5.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyee==13.0.0
pylibsrtp==0.12.0
pyloudnorm==0.1.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
pytest==8.3.5
python-dotenv==1.1.0
requests==2.32.3
resampy==0.4.3
rsa==4.9.1
ruff==0.11.9
scipy==1.15.3
sniffio==1.3.1
soxr==0.5.0.post1
sympy==1.14.0
timecode==1.4.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
websockets==13.1
yarl==1.20.0
